Moved cifer-10 to a new benchmark folder as CIFAR is not an SDM example. Updated the Readme
