+++ .gitignore
@@ -24,6 +24,9 @@ examples/inference/sentinel-2a-rgbnir/outputs/*
 
 examples/kaggle/geolifeclef2022/dataset
 
+examples/benchmark/cifar-10/dataset/*
+examples/benchmark/cifar-10/outputs/*
+
 # Packaging
 dist/
 build/
