--- examples/ecologists/cifar-10/outputs/cnn_cifar10/30_epochs_no_pretrained/cnn_cifar10_torchgeo.log
@@ -1,162 +1,0 @@
-[2024-04-25 16:31:11,401][malpolon][INFO] - 
-# Model specification
-[2024-04-25 16:31:11,402][malpolon][INFO] - ResNet(
-  (conv1): Conv2d(3, 64, kernel_size=(7, 7), stride=(2, 2), padding=(3, 3), bias=False)
-  (bn1): BatchNorm2d(64, eps=1e-05, momentum=0.1, affine=True, track_running_stats=True)
-  (act1): ReLU(inplace=True)
-  (maxpool): MaxPool2d(kernel_size=3, stride=2, padding=1, dilation=1, ceil_mode=False)
-  (layer1): Sequential(
-    (0): BasicBlock(
-      (conv1): Conv2d(64, 64, kernel_size=(3, 3), stride=(1, 1), padding=(1, 1), bias=False)
-      (bn1): BatchNorm2d(64, eps=1e-05, momentum=0.1, affine=True, track_running_stats=True)
-      (drop_block): Identity()
-      (act1): ReLU(inplace=True)
-      (aa): Identity()
-      (conv2): Conv2d(64, 64, kernel_size=(3, 3), stride=(1, 1), padding=(1, 1), bias=False)
-      (bn2): BatchNorm2d(64, eps=1e-05, momentum=0.1, affine=True, track_running_stats=True)
-      (act2): ReLU(inplace=True)
-    )
-    (1): BasicBlock(
-      (conv1): Conv2d(64, 64, kernel_size=(3, 3), stride=(1, 1), padding=(1, 1), bias=False)
-      (bn1): BatchNorm2d(64, eps=1e-05, momentum=0.1, affine=True, track_running_stats=True)
-      (drop_block): Identity()
-      (act1): ReLU(inplace=True)
-      (aa): Identity()
-      (conv2): Conv2d(64, 64, kernel_size=(3, 3), stride=(1, 1), padding=(1, 1), bias=False)
-      (bn2): BatchNorm2d(64, eps=1e-05, momentum=0.1, affine=True, track_running_stats=True)
-      (act2): ReLU(inplace=True)
-    )
-  )
-  (layer2): Sequential(
-    (0): BasicBlock(
-      (conv1): Conv2d(64, 128, kernel_size=(3, 3), stride=(2, 2), padding=(1, 1), bias=False)
-      (bn1): BatchNorm2d(128, eps=1e-05, momentum=0.1, affine=True, track_running_stats=True)
-      (drop_block): Identity()
-      (act1): ReLU(inplace=True)
-      (aa): Identity()
-      (conv2): Conv2d(128, 128, kernel_size=(3, 3), stride=(1, 1), padding=(1, 1), bias=False)
-      (bn2): BatchNorm2d(128, eps=1e-05, momentum=0.1, affine=True, track_running_stats=True)
-      (act2): ReLU(inplace=True)
-      (downsample): Sequential(
-        (0): Conv2d(64, 128, kernel_size=(1, 1), stride=(2, 2), bias=False)
-        (1): BatchNorm2d(128, eps=1e-05, momentum=0.1, affine=True, track_running_stats=True)
-      )
-    )
-    (1): BasicBlock(
-      (conv1): Conv2d(128, 128, kernel_size=(3, 3), stride=(1, 1), padding=(1, 1), bias=False)
-      (bn1): BatchNorm2d(128, eps=1e-05, momentum=0.1, affine=True, track_running_stats=True)
-      (drop_block): Identity()
-      (act1): ReLU(inplace=True)
-      (aa): Identity()
-      (conv2): Conv2d(128, 128, kernel_size=(3, 3), stride=(1, 1), padding=(1, 1), bias=False)
-      (bn2): BatchNorm2d(128, eps=1e-05, momentum=0.1, affine=True, track_running_stats=True)
-      (act2): ReLU(inplace=True)
-    )
-  )
-  (layer3): Sequential(
-    (0): BasicBlock(
-      (conv1): Conv2d(128, 256, kernel_size=(3, 3), stride=(2, 2), padding=(1, 1), bias=False)
-      (bn1): BatchNorm2d(256, eps=1e-05, momentum=0.1, affine=True, track_running_stats=True)
-      (drop_block): Identity()
-      (act1): ReLU(inplace=True)
-      (aa): Identity()
-      (conv2): Conv2d(256, 256, kernel_size=(3, 3), stride=(1, 1), padding=(1, 1), bias=False)
-      (bn2): BatchNorm2d(256, eps=1e-05, momentum=0.1, affine=True, track_running_stats=True)
-      (act2): ReLU(inplace=True)
-      (downsample): Sequential(
-        (0): Conv2d(128, 256, kernel_size=(1, 1), stride=(2, 2), bias=False)
-        (1): BatchNorm2d(256, eps=1e-05, momentum=0.1, affine=True, track_running_stats=True)
-      )
-    )
-    (1): BasicBlock(
-      (conv1): Conv2d(256, 256, kernel_size=(3, 3), stride=(1, 1), padding=(1, 1), bias=False)
-      (bn1): BatchNorm2d(256, eps=1e-05, momentum=0.1, affine=True, track_running_stats=True)
-      (drop_block): Identity()
-      (act1): ReLU(inplace=True)
-      (aa): Identity()
-      (conv2): Conv2d(256, 256, kernel_size=(3, 3), stride=(1, 1), padding=(1, 1), bias=False)
-      (bn2): BatchNorm2d(256, eps=1e-05, momentum=0.1, affine=True, track_running_stats=True)
-      (act2): ReLU(inplace=True)
-    )
-  )
-  (layer4): Sequential(
-    (0): BasicBlock(
-      (conv1): Conv2d(256, 512, kernel_size=(3, 3), stride=(2, 2), padding=(1, 1), bias=False)
-      (bn1): BatchNorm2d(512, eps=1e-05, momentum=0.1, affine=True, track_running_stats=True)
-      (drop_block): Identity()
-      (act1): ReLU(inplace=True)
-      (aa): Identity()
-      (conv2): Conv2d(512, 512, kernel_size=(3, 3), stride=(1, 1), padding=(1, 1), bias=False)
-      (bn2): BatchNorm2d(512, eps=1e-05, momentum=0.1, affine=True, track_running_stats=True)
-      (act2): ReLU(inplace=True)
-      (downsample): Sequential(
-        (0): Conv2d(256, 512, kernel_size=(1, 1), stride=(2, 2), bias=False)
-        (1): BatchNorm2d(512, eps=1e-05, momentum=0.1, affine=True, track_running_stats=True)
-      )
-    )
-    (1): BasicBlock(
-      (conv1): Conv2d(512, 512, kernel_size=(3, 3), stride=(1, 1), padding=(1, 1), bias=False)
-      (bn1): BatchNorm2d(512, eps=1e-05, momentum=0.1, affine=True, track_running_stats=True)
-      (drop_block): Identity()
-      (act1): ReLU(inplace=True)
-      (aa): Identity()
-      (conv2): Conv2d(512, 512, kernel_size=(3, 3), stride=(1, 1), padding=(1, 1), bias=False)
-      (bn2): BatchNorm2d(512, eps=1e-05, momentum=0.1, affine=True, track_running_stats=True)
-      (act2): ReLU(inplace=True)
-    )
-  )
-  (global_pool): SelectAdaptivePool2d (pool_type=avg, flatten=Flatten(start_dim=1, end_dim=-1))
-  (fc): Linear(in_features=512, out_features=10, bias=True)
-)
-[2024-04-25 16:31:11,406][malpolon][INFO] - CrossEntropyLoss()
-[2024-04-25 16:31:11,406][malpolon][INFO] - SGD (
-Parameter Group 0
-    dampening: 0
-    differentiable: False
-    foreach: None
-    lr: 0.01
-    maximize: False
-    momentum: 0.9
-    nesterov: True
-    weight_decay: 0
-)
-[2024-04-25 16:31:11,407][malpolon][INFO] - Metrics: {'multiclass_accuracy': {'kwargs': {'num_classes': 10}, 'callable': <function multiclass_accuracy at 0x730ce3065c60>}}
-[2024-04-25 16:31:11,407][malpolon][INFO] - 
-# Data loading information
-[2024-04-25 16:31:11,407][malpolon][INFO] - 
-## Training data
-[2024-04-25 16:31:11,407][malpolon][INFO] - Train dataset: Dataset CIFAR10
-    Number of datapoints: 50000
-    Root location: dataset/cifar_from_torchvision
-    Split: Train
-    StandardTransform
-Transform: ToTensor()
-[2024-04-25 16:31:11,407][malpolon][INFO] - Train set size: 50000
-[2024-04-25 16:31:11,407][malpolon][INFO] - Train data transformations: ToTensor()
-[2024-04-25 16:31:11,407][malpolon][INFO] - Train data target transformations: None
-[2024-04-25 16:31:11,407][malpolon][INFO] - Train data sampler: SequentialSampler(
-    data_source=<object>
-)
-[2024-04-25 16:31:11,407][malpolon][INFO] - Train data batch sampler: BatchSampler(
-    sampler=<object>, batch_size=32, drop_last=False
-)
-[2024-04-25 16:31:11,407][malpolon][INFO] - 
-## Validation data
-[2024-04-25 16:31:11,408][malpolon][INFO] - Validation dataset: Dataset CIFAR10
-    Number of datapoints: 10000
-    Root location: dataset/cifar_from_torchvision
-    Split: Test
-    StandardTransform
-Transform: ToTensor()
-[2024-04-25 16:31:11,408][malpolon][INFO] - Validation set size: 10000
-[2024-04-25 16:31:11,408][malpolon][INFO] - Validation data transformations: ToTensor()
-[2024-04-25 16:31:11,408][malpolon][INFO] - Validation data target transformations: None
-[2024-04-25 16:31:11,408][malpolon][INFO] - Validation data sampler: SequentialSampler(
-    data_source=<object>
-)
-[2024-04-25 16:31:11,408][malpolon][INFO] - Validation data batch sampler: BatchSampler(
-    sampler=<object>, batch_size=32, drop_last=False
-)
-[2024-04-25 16:31:11,408][malpolon][INFO] - 
-# Strategy information
-[2024-04-25 16:31:11,408][malpolon][INFO] - <pytorch_lightning.strategies.single_device.SingleDeviceStrategy object at 0x730c5f61a560>
