--- examples/ecologists/cifar-10/outputs/cnn_cifar10/30_epochs_no_pretrained/scores_test_dataset.csv
@@ -1,2 +1,0 @@
-metric_name,score
-multiclass_accuracy,0.7539000511169434
