--- examples/ecologists/cifar-10/outputs/cnn_cifar10/30_epochs_pretrained/.hydra/overrides.yaml
@@ -1,1 +1,0 @@
-[]
